+++ src/main/java/com/ya/pf/auditable/customer/CustomerServiceImpl.java
@@ -72,7 +72,7 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
-    public void updateCustomerBalance(long customerId, double customerBalance) {
+    public void updateCustomerBalance(long customerId, float customerBalance) {
 
         customerRepository.updateCustomerBalance(customerId, customerBalance);
     }
