+++ src/main/java/com/ya/pf/auditable/customer/dto/CustomerDTO.java
@@ -1,5 +1,5 @@
 package com.ya.pf.auditable.customer.dto;
 
-public record CustomerDTO(long id, String name, Double balance) {
+public record CustomerDTO(long id, String name, Float balance) {
 
 }
