+++ src/main/java/com/ya/pf/auditable/supplier/dto/SupplierDTO.java
@@ -1,5 +1,5 @@
 package com.ya.pf.auditable.supplier.dto;
 
-public record SupplierDTO(long id, String name, Double balance) {
+public record SupplierDTO(long id, String name, Float balance) {
 
 }
